+++ apps/web/.next/server/server-reference-manifest.js
@@ -1,1 +1,1 @@
-self.__RSC_SERVER_MANIFEST="{\n  \"node\": {},\n  \"edge\": {},\n  \"encryptionKey\": \"zFVWi4Gx5/8Mup3wFmjFyltnpc3phgufCsNS1m0we0A=\"\n}"
+self.__RSC_SERVER_MANIFEST="{\n  \"node\": {},\n  \"edge\": {},\n  \"encryptionKey\": \"gtVKU8N2+BAEAB4yDveGlaDeuR1t2O3Li4wfwmzka3M=\"\n}"